refactor(migration): dooboo-cli renewal (#92)
We're thrilled to announce the latest changes to dooboo-cli! We have decided to dedicate our efforts to bolster the development of our boilerplate with an exclusive focus on Expo app using Expo router. This means other boilerplates will no longer receive regular updates, as we believe concentrating on the latest technologies is key to staying relevant and providing the best tooling for developers.

One of the exciting updates is the newly added support for `npx`, a package runner tool that comes with `npm`. With this update, you no longer need to install `dooboo-cli` globally. You can now simply run `dooboo-cli` by typing `npx dooboo` in your terminal. 

We are convinced that these changes will streamline your development process and enhance your experience with dooboo-cli. Stay tuned for more exciting updates!